% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{chart_performance_summary}
\alias{chart_performance_summary}
\title{chart_performance_summary}
\usage{
chart_performance_summary(rets_xts, na_fill = FALSE, ...)
}
\arguments{
\item{rets_xts}{xts, an xts of returns}

\item{na_fill}{logical, if TRUE, NAs will be changed to zeros, Default: FALSE}

\item{...}{additional parameters passed to \code{PerformanceAnalytics::charts.PerformanceSummary()}}
}
\value{
See documentation for \code{PerformanceAnalytics::charts.PerformanceSummary()}
}
\description{
A wrapper for \code{PerformanceAnalytics::charts.PerformanceSummary()}.
}
