% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{chart_risk_return_scatter}
\alias{chart_risk_return_scatter}
\title{chart_risk_return_scatter}
\usage{
chart_risk_return_scatter(rets_xts, ...)
}
\arguments{
\item{rets_xts}{xts, an xts of returns}

\item{...}{additional parameters passed to \code{PerformanceAnalytics::chart.RiskReturnScatter()}}
}
\value{
See documentation for \code{PerformanceAnalytics::chart.RiskReturnScatter()}
}
\description{
A wrapper for \code{PerformanceAnalytics::chart.RiskReturnScatter()}.
}
