% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{chart_rolling_performance}
\alias{chart_rolling_performance}
\title{chart_rolling_performance}
\usage{
chart_rolling_performance(rets_xts, ...)
}
\arguments{
\item{rets_xts}{xts, an xts of returns}

\item{...}{additional parameters passed to \code{PerformanceAnalytics::charts.RollingPerformance()}}
}
\value{
See documentation for \code{PerformanceAnalytics::charts.RollingPerformance()}
}
\description{
A wrapper for \code{PerformanceAnalytics::charts.RollingPerformance()}.
}
