% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_permutation_data.R
\name{make_permutation_data}
\alias{make_permutation_data}
\title{make_permutation_data}
\usage{
make_permutation_data(
  model_funs,
  model_units,
  max_units,
  make_contributions = TRUE,
  benchmark_symbol = "SPY"
)
}
\arguments{
\item{model_funs}{a named vector of model functions}

\item{model_units}{numeric, a vector of units to allocate to a certain model (the ordering of this vector corresponds with the ordering of the \code{model_funs} parameter)}

\item{max_units}{numeric, the maximum number of units which may be allocated to the portfolio}

\item{make_contributions}{logical, if TRUE, the portfolio contributions will be outputted (better performance if set to FALSE), Default: TRUE}

\item{benchmark_symbol}{character, the symbol whose return series will be used as a benchmark, Default: 'SPY'}
}
\value{
A list of data points, including model/benchmark returns, (optionally) model/asset portfolio contributions, correlations, and ancillary information.
}
\description{
Provides an object which can then be passed to other functions to access the viability of each permutation of a collection of models.
}
\details{
Before use, make sure that \code{model_funs} is named, and that the necessary OHLCV series are in \code{./prices}.  This function is designed to create an object to be passed to other functions.  See documentation in the \emph{References} section (below).
}
\examples{
\dontrun{
if(interactive()){
 # Where the units for CAN is 1.5, Gan: 1, and KDA: 0.5
 perm_data <- make_permutation_data(
   model_funs = c(
     CAN = clmodels::CAN, 
     GAN = clmodels::GAN,
     KDA = clmodels::KDA),
   model_units = c(1.5, 1, 0.5),
   max_units = 3,
   make_contributions = TRUE,
   benchmark_symbol = 'SPY'
 )
 }
}
}
\references{
\url{https://github.com/causality-loop/cltest}
}
