% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_report_xts.R
\name{make_report_xts}
\alias{make_report_xts}
\title{make_report_xts}
\usage{
make_report_xts(make_permutation_data_output, walk_forward_test_output)
}
\arguments{
\item{make_permutation_data_output}{the output of the \code{make_permutation_data()} function}

\item{walk_forward_test_output}{the output of the \code{walk_forward_test()} function}
}
\value{
An xts of model returns at various lookback year lengths (with benchmark return series for comparison).
}
\description{
Makes an xts which is later passed to other functions which create tables and plots for walk-forward test performance reporting.
}
\details{
See \emph{Examples} for usage.  Takes output from \code{make_permutation_data()} and \code{walk_forward_test()} to create an xts comparing the return series of each walk-forward test using different lookback year lengths to the returns of a user-specified benchmark.
}
\examples{
\dontrun{
if(interactive()){
 perm_data <- make_permutation_data(
   model_funs = c(
     CAN = clmodels::CAN, 
     GAN = clmodels::GAN,
     KDA = clmodels::KDA),
   model_units = c(1.5, 1, 0.5),
   max_units = 3,
   make_contributions = TRUE,
   benchmark_symbol = 'SPY'
 )
 wftest_list <- walk_forward_test(perm_data)
 report_xts <- make_report_xts(perm_data, wftest_list)
 }
}
}
