% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{table_annualized_returns}
\alias{table_annualized_returns}
\title{table_annualized_returns}
\usage{
table_annualized_returns(rets_xts, ...)
}
\arguments{
\item{rets_xts}{xts, an xts of returns}

\item{...}{additional parameters passed to \code{PerformanceAnalytics::table.AnnualizedReturns()}}
}
\value{
See documentation for \code{PerformanceAnalytics::table.AnnualizedReturns()}.
}
\description{
A wrapper for \code{PerformanceAnalytics::table.AnnualizedReturns}.
}
