% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{table_calendar_returns}
\alias{table_calendar_returns}
\title{table_calendar_returns}
\usage{
table_calendar_returns(rets_xts, ...)
}
\arguments{
\item{rets_xts}{xts, an xts of returns}

\item{...}{additional parameters passed to \code{PerformanceAnalytics::table.CalendarReturns()}}
}
\value{
A list of calendar returns.
}
\description{
Makes a list of calendar returns.
}
