% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{table_contribution}
\alias{table_contribution}
\title{table_contribution}
\usage{
table_contribution(make_permutation_data_output)
}
\arguments{
\item{make_permutation_data_output}{the output of the \code{make_permutation_data()} function}
}
\value{
A list of data tables.
}
\description{
Makes a list of data tables indicating the portfolio contribution for each model, asset.
}
