% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{table_drawdowns}
\alias{table_drawdowns}
\title{table_drawdowns}
\usage{
table_drawdowns(rets_xts, ...)
}
\arguments{
\item{rets_xts}{xts, an xts of returns}

\item{...}{additional parameters passed to \code{PerformanceAnalytics::table.Drawdowns()}}
}
\value{
See documentation for \code{PerformanceAnalytics::table.Drawdowns()}.
}
\description{
A wrapper for \code{PerformanceAnalytics::table.Drawdowns}.
}
