% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walk_forward_test.R
\name{walk_forward_test}
\alias{walk_forward_test}
\title{walk_forward_test}
\usage{
walk_forward_test(
  make_permutation_data_output,
  lookback_year_seq = c(1, 3, 5, 10)
)
}
\arguments{
\item{make_permutation_data_output}{the output of the \code{make_permutation_data()} function}

\item{lookback_year_seq}{integer, a sequence of integers indicating various lookback lengths (in years), Default: c(1, 3, 5, 10)}
}
\value{
A list of returns (in xts format), ancillary information for each lookback length, as well as the user-inputted \code{lookback_year_seq}.
}
\description{
At quarterly intervals, select the best model portfolio permutation using a training set of lookback length = X years.
}
\details{
Designed to be passed to the \code{make_report_xts()} function.
}
